Do not save eligibility criteria in DB
