+++ itou/templates/apply/process_eligibility.html
@@ -14,26 +14,23 @@
 
         {% bootstrap_form_errors form %}
 
-        {% include "eligibility/includes/form_v_1_0_0.html" with form=form %}
+        {% include "eligibility/includes/form.html" with form=form %}
 
-        <p>
-            <a class="btn btn-secondary" href="{% url 'apply:details_for_siae' job_application_id=job_application.id %}">{% trans "Annuler" %}</a>
-        </p>
-
-        {% if form.errors %}
-
-            <div class="alert alert-secondary" role="alert">
-                <p>
-                    {% trans "Si le candidat n'est pas éligible à l'insertion par l'activité économique, vous pouvez décliner sa candidature :" %}
-                </p>
-                <a href="{% url 'apply:refuse' job_application_id=job_application.id %}" class="btn btn-outline-danger btn-block">
-                    {% trans "Je ne veux pas l'embaucher" %}
-                    {% include "includes/icon.html" with icon="arrow-right" %}
-                </a>
-            </div>
-
-        {% endif %}
+        <div class="alert alert-secondary" role="alert">
+            <p>
+                {% trans "Si le candidat n'est pas éligible à l'insertion par l'activité économique, vous pouvez décliner sa candidature :" %}
+            </p>
+            <a href="{% url 'apply:refuse' job_application_id=job_application.id %}" class="btn btn-outline-danger btn-block">
+                {% trans "Je ne veux pas l'embaucher" %}
+                {% include "includes/icon.html" with icon="arrow-right" %}
+            </a>
+        </div>
 
     </form>
+
+    <hr>
+    <p>
+        <a href="{% url 'apply:details_for_siae' job_application_id=job_application.id %}">{% trans "Annuler" %}</a>
+    </p>
 
 {% endblock %}
