+++ itou/templates/apply/submit_step_eligibility.html
@@ -14,7 +14,7 @@
 
         {% bootstrap_form_errors form %}
 
-        {% include "eligibility/includes/form_v_1_0_0.html" with form=form %}
+        {% include "eligibility/includes/form.html" with form=form %}
 
     </form>
