--- itou/templates/eligibility/includes/form_v_1_0_0.html
@@ -1,96 +1,0 @@
-{% load i18n %}
-{% load bootstrap4 %}
-{% load dict_lookup %}
-
-<div class="alert alert-warning" role="alert">
-    {% if request.user.is_prescriber %}
-        {% trans "Pour valider une éligibilité vous devez indiquer au moins un besoin d'accompagnement." %}
-    {% endif %}
-    {% if request.user.is_siae_staff %}
-        <p>{% trans "Pour valider une éligibilité vous devez :" %}</p>
-        <ul class="mb-0">
-            <li>{% trans "indiquer au moins un besoin d'accompagnement" %}</li>
-            <li><b>{% trans "et" %}</b></li>
-            <li>
-                {% trans "indiquer des critères administratifs :" %}
-                <ul>
-                    <li>{% trans "au moins un critère administratif de niveau 1" %}</li>
-                    <li><b>{% trans "ou" %}</b></li>
-                    <li>{% trans "le cumul d'au moins trois critères de niveau 2" %}</li>
-                </ul>
-            </li>
-        </ul>
-    {% endif %}
-</div>
-
-{% for field in form %}
-
-    {# TODO: find an elegant way to display the category titles in template. #}
-
-    {% if field.name == "faire_face_a_des_difficultes_administratives_ou_juridiques" %}
-        <hr>
-        <h2><span class="badge badge-secondary">{% trans "Besoins d'accompagnement" %}</span></h2>
-        <hr>
-    {% endif %}
-
-    {% if field.name == "criteres_administratifs_de_niveau_1" %}
-        <hr>
-        <h2><span class="badge badge-secondary">{% trans "Critères administratifs" %}</span></h2>
-        <hr>
-    {% endif %}
-
-    {% if field.field.widget.allow_multiple_selected %}
-        <div class="form-group">
-            <h4>{{ field.label }}</h4>
-            <ul class="list-unstyled">
-                {% for radio in field %}
-                    <li>
-                        {% with form.extra_info|dict_lookup:radio.data.value as info %}
-                            <div class="clearfix mb-2">
-                                <div class="float-left">
-                                    {{ radio.tag }}
-                                </div>
-                                <div class="pl-4">
-                                    <label for="{{ radio.id_for_label }}" class="mb-0">
-                                        {{ radio.choice_label }}
-                                        {% if info.help %}
-                                            <br>
-                                            <small class="text-muted">{{ info.help }}</small>
-                                        {% endif %}
-                                    </label>
-                                    {% if info.written_proof %}
-                                        <br>
-                                        <small>
-                                            {% trans "Pièce justificative :" %}
-                                            <i>{{ info.written_proof }}</i>
-                                        </small>
-                                    {% endif %}
-                                    {% if info.url %}
-                                        <br>
-                                        <small>
-                                            <a href="{{ info.url }}" rel="noopener" target="_blank">
-                                                {{ info.url }}
-                                            </a>
-                                        </small>
-                                    {% endif %}
-                                </div>
-                            </div>
-                        {% endwith %}
-                    </li>
-                {% endfor %}
-            </ul>
-        </div>
-    {% endif %}
-
-{% endfor %}
-
-<hr>
-
-<div class="alert alert-warning" role="alert">
-    <p>
-        {% blocktrans %}En cliquant sur "Valider" vous confirmez que la personne concernée a donné son consentement exprès (démarche active, explicite et de préférence écrite, qui doit être libre, spécifique, et informée).{% endblocktrans %}
-    </p>
-    {% buttons %}
-        <button type="submit" class="btn btn-primary">{% trans "Valider" %}</button>
-    {% endbuttons %}
-</div>
